[MOD/#130] 뒤로가기 화면 전환 오류
