+++ app/src/main/java/com/el/yello/presentation/tutorial/TutorialCActivity.kt
@@ -1,18 +1,18 @@
-package com.el.yello.presentation.tuorial
+package com.el.yello.presentation.tutorial
 
 import android.content.Intent
 import android.os.Bundle
 import com.el.yello.R
-import com.el.yello.databinding.ActivityTutorialABinding
+import com.el.yello.databinding.ActivityTutorialCBinding
 import com.example.ui.base.BindingActivity
 import com.example.ui.view.setOnSingleClickListener
 
-class TutorialAActivity : BindingActivity<ActivityTutorialABinding>(R.layout.activity_tutorial_a) {
+class TutorialCActivity : BindingActivity<ActivityTutorialCBinding>(R.layout.activity_tutorial_c) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding.root.setOnSingleClickListener {
-            val intent = Intent(this, TutorialBActivity::class.java)
+            val intent = Intent(this, TutorialDActivity::class.java)
             startActivity(intent)
         }
     }
