+++ app/src/main/java/com/el/yello/presentation/tutorial/TutorialEndPlusActivity.kt
@@ -1,4 +1,4 @@
-package com.el.yello.presentation.tuorial
+package com.el.yello.presentation.tutorial
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,8 +8,8 @@ import com.el.yello.presentation.main.MainActivity
 import com.example.ui.base.BindingActivity
 import com.example.ui.view.setOnSingleClickListener
 
-class TutorialEndActivity :
-    BindingActivity<ActivityTutorialEndPointBinding>(R.layout.activity_tutorial_end_point) {
+class TutorialEndPlusActivity :
+    BindingActivity<ActivityTutorialEndPointBinding>(R.layout.activity_tutorial_end_pluspoint) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
